% Auto-generated: do not edit by hand
\name{featureTable}

\alias{featureTable}

\title{FeatureTable component}

\description{
FeatureTable Component Displays feature data in a scrollable table with dynamic columns and interaction
}

\usage{
featureTable(contributions=NULL, data=NULL, height=NULL, hoveredId=NULL,
idColumn=NULL, onClick=NULL, onHover=NULL, onScroll=NULL,
style=NULL)
}

\arguments{
\item{contributions}{}

\item{data}{List of named lists. }

\item{height}{Numeric. }

\item{hoveredId}{Character. }

\item{idColumn}{Character. }

\item{onClick}{}

\item{onHover}{}

\item{onScroll}{}

\item{style}{Lists containing elements 'textcolor', 'background', 'headerbackground', 'highlightbackground'.
those elements have the following types:
  - textcolor (character; optional)
  - background (character; optional)
  - headerbackground (character; optional)
  - highlightbackground (character; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

